fix: move serializer folder to src
