+++ src/utils/index.ts
@@ -2,12 +2,4 @@ export { capitalize } from "./generics";
 
 export { Puppeteer } from "./puppeteer";
 
-export {
-  matchSerializer,
-  playerMatchSerializer,
-  scoreboardSerializer,
-  platformCredentialsSerializer,
-  playerSerializer,
-} from "./serializers";
-
 export { pageOr404, validateAndReturnUrlAndId } from "./services";
